+++ src/main/kotlin/onku/backend/domain/notice/controller/user/NoticeController.kt
@@ -1,12 +1,12 @@
-package onku.backend.domain.notice.controller
+package onku.backend.domain.notice.controller.user
 
 import io.swagger.v3.oas.annotations.Operation
 import io.swagger.v3.oas.annotations.tags.Tag
-import jakarta.validation.Valid
 import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.*
 import onku.backend.domain.member.Member
-import onku.backend.domain.notice.dto.notice.*
+import onku.backend.domain.notice.dto.notice.NoticeDetailResponse
+import onku.backend.domain.notice.dto.notice.NoticeListItemResponse
 import onku.backend.domain.notice.service.NoticeService
 import onku.backend.global.annotation.CurrentMember
 import onku.backend.global.page.PageResponse
@@ -15,8 +15,8 @@ import onku.backend.global.response.SuccessResponse
 @RestController
 @RequestMapping("/api/v1/notice")
 @Tag(
-    name = "[STAFF] 공지 API",
-    description = "공지 관련 API"
+    name = "[USER/MANAGEMENT] 공지 API",
+    description = "공지 조회/검색 API"
 )
 class NoticeController(
     private val noticeService: NoticeService
@@ -24,7 +24,7 @@ class NoticeController(
 
     @GetMapping
     @Operation(
-        summary = "공지 리스트 조회 [운영진]",
+        summary = "공지 리스트 조회",
         description = "id, 제목, 작성자id, 작성자이름, {카테고리 이름/색}, 작성일(YYYY/MM/DD HH:MM), 상태, 이미지, 파일을 페이징하여 반환. 전체 조회가 필요한 경우 카테고리 id는 비워두시면 됩니다."
     )
     fun list(
@@ -40,7 +40,7 @@ class NoticeController(
 
     @GetMapping("/{noticeId}")
     @Operation(
-        summary = "공지 단일 조회 [운영진]",
+        summary = "공지 단일 조회",
         description = "제목, 카테고리, 작성일자(MM:dd HH:MM), 내용, 작성자id, 작성자이름, 이미지, 파일 presigned url 리스트"
     )
     fun get(
@@ -51,46 +51,9 @@ class NoticeController(
         return ResponseEntity.ok(SuccessResponse.ok(body))
     }
 
-    @PostMapping
-    @Operation(
-        summary = "공지 등록 [운영진]",
-        description = "제목, 카테고리, 내용, 이미지, pdf로 등록"
-    )
-    fun create(
-        @CurrentMember member: Member,
-        @RequestBody @Valid req: NoticeCreateRequest
-    ): ResponseEntity<SuccessResponse<NoticeDetailResponse>> {
-        val body = noticeService.create(member, req)
-        return ResponseEntity.ok(SuccessResponse.ok(body))
-    }
-
-    @PutMapping("/{noticeId}")
-    @Operation(
-        summary = "공지 수정 [운영진]",
-        description = "등록과 동일한 request/response"
-    )
-    fun update(
-        @PathVariable noticeId: Long,
-        @CurrentMember member: Member,
-        @RequestBody @Valid req: NoticeUpdateRequest
-    ): ResponseEntity<SuccessResponse<NoticeDetailResponse>> {
-        val body = noticeService.update(noticeId, member, req)
-        return ResponseEntity.ok(SuccessResponse.ok(body))
-    }
-
-    @DeleteMapping("/{noticeId}")
-    @Operation(
-        summary = "공지 삭제 [운영진]",
-        description = "공지 삭제"
-    )
-    fun delete(@PathVariable noticeId: Long): ResponseEntity<SuccessResponse<Unit>> {
-        noticeService.delete(noticeId)
-        return ResponseEntity.ok(SuccessResponse.ok(Unit))
-    }
-
     @GetMapping("/search")
     @Operation(
-        summary = "공지 검색 [학회원]",
+        summary = "공지 검색",
         description = "검색어로 공지 제목/내용에서 검색하여 페이징 반환합니다."
     )
     fun search(
